+++ page-register-pay.php
@@ -18,15 +18,7 @@ get_header(); ?>
 
 <h3>2⃣ Betala medlemskap</h3>
 <hr>
-<p>→ Tryck på knappen nedanför för att betala 50 kronor.</p>
-
-<h3>3⃣ Verifiera e-postadress</h3>
-<hr>
-<p>→ Kolla din inkorg och tryck på länken i det mail som vi skickat.</p>
-
-<div class="wpum-message warning">
-<p>⚠ OBS! Du måste ange samma e-postadress i betalningen som i formuläret.</p>
-</div>
+<p>→ Tryck på knappen för att betala 50 kronor.</p>
 
 <?php
 // Stripe Sandbox?
@@ -38,8 +30,18 @@ if ($test_mode) {
 }
 ?>
 
-<p><button type="submit"><a href="<?php echo esc_url($payment_link); ?>">💳 Betala 50 kr</a></button></p>
+<p><button class="green" onclick="window.location.href='<?php echo esc_url($payment_link); ?>'">💳 Betala 50 kr</button></p>
 <p class="info">💡 Problem eller frågor? Maila <a href="mailto:info@loopis.org">info@loopis.org</a></p>
+
+<div class="wpum-message warning">
+<p>⚠ OBS! Du måste ange samma e-postadress i betalningen som i formuläret.</p>
+</div>
+
+<p><span class="big-link"><a href="<?php echo esc_url(add_query_arg('option', 'swish-membership', home_url('/shop/'))); ?>">💸 Betala med Swish istället</a></span></p>
+
+<h3>3⃣ Verifiera e-postadress</h3>
+<hr>
+<p>→ Kolla din inkorg och tryck på länken i det mail som vi skickat.</p>
 
 <?php
 // Developer info in test mode