+++ page-register-welcome.php
@@ -28,6 +28,6 @@ get_header(); ?>
 <p><strong>→ Har du tryckt på länken i det mail vi skickat? Då kan du logga in!</strong></p>
 
 <?php insert_spacer(20); ?>
-/* What the heck? */
+
 <p><button type="submit"><a href="<?php echo esc_url(home_url('/'));?>">Logga in!</a></button></p>
 <p class="info">💡 Problem eller frågor? Maila <a href="mailto:info@loopis.org">info@loopis.org</a></p>